moved workaround logic
